requirements.txt
keras
tensorflow-gpu
onnx
h5py
hdf5plugin
onnx2keras